Added CreateConstraints command and handler
