+++ profile/src/main/java/com/scottlogic/deg/profile/ProfileCommandBus.java
@@ -6,11 +6,13 @@ import com.google.inject.Inject;
 import com.scottlogic.deg.common.commands.CommandBase;
 import com.scottlogic.deg.common.commands.CommandBus;
 import com.scottlogic.deg.common.commands.CommandResult;
-import com.scottlogic.deg.profile.handlers.CreateFieldsHandler;
-import com.scottlogic.deg.profile.handlers.CreateProfileHandler;
+import com.scottlogic.deg.profile.reader.handlers.CreateConstraintsHandler;
+import com.scottlogic.deg.profile.reader.handlers.CreateFieldsHandler;
+import com.scottlogic.deg.profile.reader.handlers.CreateProfileHandler;
 import com.scottlogic.deg.profile.reader.FileReader;
-import com.scottlogic.deg.profile.validators.CreateFieldsValidator;
-import com.scottlogic.deg.profile.validators.CreateProfileValidator;
+import com.scottlogic.deg.profile.reader.validators.CreateConstraintsValidator;
+import com.scottlogic.deg.profile.reader.validators.CreateFieldsValidator;
+import com.scottlogic.deg.profile.reader.validators.CreateProfileValidator;
 
 import java.util.stream.Stream;
 
@@ -22,8 +24,9 @@ public class ProfileCommandBus implements CommandBus
     public ProfileCommandBus(FileReader fileReader)
     {
         pipeline = new Pipelinr(() -> Stream.of(
-            new CreateProfileHandler(fileReader, this, new CreateProfileValidator()),
-            new CreateFieldsHandler(new CreateFieldsValidator())));
+            new CreateProfileHandler(this, new CreateProfileValidator()),
+            new CreateFieldsHandler(new CreateFieldsValidator()),
+            new CreateConstraintsHandler(fileReader, new CreateConstraintsValidator())));
     }
 
     @Override
