+++ profile/src/main/java/com/scottlogic/deg/profile/reader/validators/CreateFieldsValidator.java
@@ -1,8 +1,8 @@
-package com.scottlogic.deg.profile.validators;
+package com.scottlogic.deg.profile.reader.validators;
 
 import com.scottlogic.deg.common.validators.ValidationResult;
 import com.scottlogic.deg.common.validators.Validator;
-import com.scottlogic.deg.profile.commands.CreateFields;
+import com.scottlogic.deg.profile.reader.commands.CreateFields;
 
 public class CreateFieldsValidator implements Validator<CreateFields>
 {
